refactor: Cart functionality with CartMain, CartDrawer, and CartLoading components for improved shopping experience
