+++ app/routes/($locale).cart.tsx
@@ -3,9 +3,9 @@ import {Suspense} from 'react';
 import type {CartQueryDataReturn} from '@shopify/hydrogen';
 import {CartForm} from '@shopify/hydrogen';
 import {json, type ActionFunctionArgs} from '@shopify/remix-oxygen';
-import {CartMain} from '~/components/Cart';
 import {useRootLoaderData} from '~/root';
 import { CartApiQueryFragment } from 'storefrontapi.generated';
+import { CartMain } from '~/components/cart/Cart';
 
 export const meta: MetaFunction = () => {
   return [{title: `Hydrogen | Cart`}];
